+++ sdk/appcenter/src/main/java/com/microsoft/appcenter/utils/NetworkStateHelper.java
@@ -176,11 +176,7 @@ public class NetworkStateHelper implements Closeable {
 
         /*
          * We will have WIFI network available event before we lose mobile network.
-         * Pending calls just before the switch might take a while to fail.
-         * When we lose a network, but have another one available, just simulate network
-         * down then up again to properly reset pending calls so that they are reliable
-         * and fast. This notification scheme is similar to the old connectivity receiver
-         * implementation.
+         * This notification scheme is similar to the old connectivity receiver implementation.
          */
         AppCenterLog.debug(LOG_TAG, "Network lost netId: " + network);
         mAvailableNetworks.remove(network);
@@ -218,10 +214,6 @@ public class NetworkStateHelper implements Closeable {
          * This code is used to notify listeners only when the network state goes from
          * connected to disconnected and vice versa
          * (without duplicate calls, the sequence will be consistent).
-         *
-         * If we switch from WIFI to Mobile and vice versa,
-         * it can take a while for pending network calls to fail because of that.
-         * We'll simulate a network state down event to the listeners to help with that scenario.
          */
         String previousNetworkType = mNetworkType;
         updateNetworkType();